requirements.txt
# Core dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.7.0
pytest-asyncio>=0.23.0

# Code quality
flake8>=6.1.0
black>=23.12.0
isort>=5.13.0

# TUI dependencies
textual>=0.50.0
rich>=13.0.0
pyyaml>=6.0.0

# Build dependencies
pyinstaller>=6.0.0
